requirements.txt
# Core Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
transformers>=4.30.0

# LangChain and LangGraph for Agent Orchestration
langchain>=0.1.0
langgraph>=0.0.30
langchain-core>=0.1.0

# Computer Vision Libraries
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
imageio>=2.31.0

# Image Processing and Enhancement
albumentations>=1.3.1
kornia>=0.7.0

# Model-Specific Libraries
diffusers>=0.21.0
accelerate>=0.23.0
safetensors>=0.3.3
timm>=0.9.7

# Metrics and Evaluation
lpips>=0.1.4
pytorch-fid>=0.3.0
scipy>=1.11.0
numpy>=1.24.0
pandas>=2.0.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing
h5py>=3.9.0
einops>=0.6.1

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# API and Async Support
aiohttp>=3.8.5
requests>=2.31.0

# File Handling
python-magic>=0.4.27
